requirements.txt
databricks-sdk==0.76.0
pandas==2.3.3
pyarrow==22.0.0
pydantic==2.12.5
pyyaml==6.0.3
requests==2.32.5
yfinance==1.0